refactors code for future unit testing
makes binding output format consistent with PCF expectations

The previous implementation of calls to the Operator required a
live connection to the Operator and Kube environment which does not
allow isolated unit testing of the broker endpoint logic without
those live environments.

This change refactors the dependency and defines a broker interface
that can be swapped between a mock implementation and the live
environment one for testing purposes.

Additionally, the binding credential format was an arbitrary JSON
structure that dumped service metadata. The binding information was
restructured to follow conventions used by OSB consumers, like those
using the Pivotal Application Service.