restructures pgo-osb internal packages
Moves both of OSB's packages into the pkg/ directory and renames them
to reflect their current usage.

What was previously named 'broker' was the bridging logic between
the OSB handler implementations and 'drivers' for executing those
requests. In this vein, the new broker interface is more correctly
called broker and the former broker library is now the osb-bridge