establishes initial base of unit test framework
integration tests can likely follow a similar pattern, but rely
on a corrected implementation of sync/async handling to know when
to check for state changes in multi-step tests

Additionally, this change replaces the flawed and abandoned satori
UUID library with the actively maintained and fixed gofrs one.
The satori library had an implementation error that drastically
reduced the uniqueness of its UUIDs.