adds support for creating and deleting users
The operator creates a set of users by default for created clusters.
However, the OSB user model uses one discardable user per service
binding. Previously, the OSB would create a binding record in the
service catalog but would not provision any users. Instead, it
returned the credentials for all of the created users.

With this change, the binding process creates a unique credential
for each binding request and binding deletions remove that unique
credential instead of doing nothing.