implements multiple plan support
This implementation starts with basic 't-shirt' sizes against
common standalone and replicated (HA) deployment topologies.

Enabling these plans requires configuration changes to the operator
in the Operator's pgo.yaml file to define the storage and container
resource definitions that govern the meaning of the 'sizes'.

Storage and Container Resource definitions should exist for:
    osbsmall - used by standalone_sm and ha_sm
    osbmedium - used by standalone_md and ha_md
    osblarge - used by standalone_lg and ha_lg

These definitions are not required if the plans that use them are
disabed through the catalog.