adds error response for deletion call while bound
The Kube Service Catalog prevents the deprovisioning of a service
if there are bound users, in compliance with the Open Service Broker
specification. However, if the OSB's deprovision endpoint is called
directly, this guarantee fails and the deletion request gets stuck
waiting for the bindings to be deleted.

This change adds a check during deprovisioning to ensure the OSB
implementation upholds the OSB specification and avoids getting stuck
on deletion without feedback.