Update Service Validation for HA Clusters
The validation that currently occurs when obtaining cluster details
during a bind operation, specifically to ensure the proper number of
services exist for the cluster, now takes into account the replica
service that will exist in HA clusters (i.e. PG clusters with one or
more replicas).  Therefore, this logic now checks the replica count
for the cluster, and if one or more replicas exists, it expects two
services to be present (one for the primary, and another for the
replicas).  If the replica count is zero, then it still only expects to see
a single service.