Disabled "all" Flag for Create Binding Show User Request
When creating binding information for a cluster, the PostgreSQL
Operator "/usershow" API endpoint is utilized to obtain user
information for the cluster.  However, while the request being sent
to this endpoint includes a selector in order to scope the users to
only the specific cluster the binding request is for, the "all" flag
is also specified, which returns users for all PG clusters.  This flag
(the "all" flag) is taking precedence over the selector specified in
the PostgreSQL Operator API Server, resulting in all users across all
PG clusters in the environment being returned when the show user
endpoint endpoint is called.  This subsequently results in errors when
binding if another PG cluster already exists in the environment, since
the binding logic will not see the expected users (i.e. it will see
users from PG clusters other than then specific cluster requested via a
selector).

This change therefore disables the "all" flag (i.e. it sets it to
false) during a bind operation to ensure only users for the specific
cluster requested (i.e. the cluster the binding request is for) are
returned.  This ensures that binding can complete properly even when
other PG clusters exist in the environment.