updates dependencies for postgres operator 4.6 line
Compatibility with the operator 4.6 line begins with 4.6.1, as 4.6.0 has
dependencies on internal packages that prevent it from being used.

This change removes support for the AuthenticateK8S option as the use
case for using the service catalog based provisioning from within
Kubernetes is very weak given the presence and dependency on the
operator. To restore this feature, the token validation logic requires
significant updates in the face of the pattern being removed from
Kubernetes over time.

To avoid surprising, breaking changes the option hasn't been removed,
but logs its lack of effect.